refactor: ParticipantRoleV2 클래스 ParticipantV2의 inner enum으로 리팩토링
