+++ src/main/java/dev/flab/studytogether/domain/studygroup/service/StudyGroupJoinService.java
@@ -3,7 +3,6 @@ package dev.flab.studytogether.domain.studygroup.service;
 import dev.flab.studytogether.domain.studygroup.entity.ParticipantV2;
 import dev.flab.studytogether.domain.studygroup.entity.StudyGroup;
 import dev.flab.studytogether.domain.studygroup.repository.StudyGroupRepository;
-import dev.flab.studytogether.domain.studygroup.role.ParticipantRoleV2;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,7 +22,7 @@ public class StudyGroupJoinService {
 
         ParticipantV2 participant = new ParticipantV2(studyGroup,
                 memberId,
-                ParticipantRoleV2.ORDINARY_PARTICIPANT,
+                ParticipantV2.ParticipantRoleV2.ORDINARY_PARTICIPANT,
                 LocalDateTime.now());
 
         studyGroup.joinGroup(participant);
