+++ src/test/java/dev/flab/studytogether/util/TestFixtureUtils.java
@@ -4,7 +4,6 @@ import dev.flab.studytogether.domain.member.entity.MemberV2;
 import dev.flab.studytogether.domain.room.entity.ActivateStatus;
 import dev.flab.studytogether.domain.studygroup.entity.ParticipantV2;
 import dev.flab.studytogether.domain.studygroup.entity.StudyGroup;
-import dev.flab.studytogether.domain.studygroup.role.ParticipantRoleV2;
 
 import java.time.LocalDateTime;
 
@@ -36,7 +35,7 @@ public class TestFixtureUtils {
 
         ParticipantV2 manager = new ParticipantV2(studyGroup,
                 1L,
-                ParticipantRoleV2.GROUP_MANAGER,
+                ParticipantV2.ParticipantRoleV2.GROUP_MANAGER,
                 LocalDateTime.of(2023, 1, 2, 4, 5));
 
         ReflectionTestUtils.setField(manager, "id", 1L);
@@ -45,7 +44,7 @@ public class TestFixtureUtils {
 
         ParticipantV2 participant = new ParticipantV2(studyGroup,
                 2L,
-                ParticipantRoleV2.ORDINARY_PARTICIPANT,
+                ParticipantV2.ParticipantRoleV2.ORDINARY_PARTICIPANT,
                 LocalDateTime.of(2023, 2, 1, 2, 3, 55));
 
         ReflectionTestUtils.setField(participant, "id", 2L);
